+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/naming/NamingSchemeAware.java
@@ -1,7 +1,0 @@
-package uk.ac.bbsrc.tgac.miso.core.service.naming;
-
-public interface NamingSchemeAware {
-
-  public void setNamingScheme(NamingScheme namingScheme);
-
-}