+++ RELEASE_NOTES.md
@@ -4,6 +4,9 @@ Changes:
 
   * Added configuration for library aliquot alias generator and validator
   * Added V2 naming scheme
+  * Added configuration for a secondary naming scheme. See the
+    [Installation Guide](https://miso-lims.readthedocs.io/projects/docs/en/latest/admin/baremetal-installation-guide/)
+    for more information
   * Fixed error deleting tissues and identities (detailed sample)
   * Fixed distributed items not being removed from boxes
   * Fixed error deleting sequencing orders